rivisto alcune classi e fixato alcuni bug

+++ src/main/java/Chems/NintendoTournaments/services/TorneoService.java
@@ -5,6 +5,7 @@ import Chems.NintendoTournaments.entities.Torneo;
 import Chems.NintendoTournaments.entities.Utente;
 import Chems.NintendoTournaments.exceptions.BadRequestException;
 import Chems.NintendoTournaments.exceptions.NotFoundException;
+import Chems.NintendoTournaments.exceptions.UnauthorizedException;
 import Chems.NintendoTournaments.payloads.TorneoDTO;
 import Chems.NintendoTournaments.repositories.TorneoRepository;
 import Chems.NintendoTournaments.repositories.GiocoRepository;
@@ -16,7 +17,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -52,7 +52,7 @@ public class TorneoService {
                 torneoDTO.statoTorneo(),
                 gioco,
                 organizzatore,
-                torneoDTO.descrizione() // Includi la descrizione
+                torneoDTO.descrizione()
         );
         return torneoRepository.save(torneo);
     }
@@ -62,25 +62,36 @@ public class TorneoService {
                 .orElseThrow(() -> new NotFoundException("Torneo non trovato con ID: " + torneoId));
     }
 
-
     public Page<Torneo> findAll(int page, int size, String sortBy) {
         Pageable pageable = PageRequest.of(page, size, Sort.by(sortBy));
         return torneoRepository.findAll(pageable);
     }
 
-    public Torneo updateTorneo(UUID torneoId, TorneoDTO torneoDTO) {
+    public Torneo updateTorneo(UUID torneoId, TorneoDTO torneoDTO, Utente utenteAutenticato) {
         Torneo found = findById(torneoId);
+
+        // Verifica se l'utente è l'organizzatore
+        if (!found.getOrganizzatore().getId().equals(utenteAutenticato.getId())) {
+            throw new UnauthorizedException("Non sei autorizzato a modificare questo torneo");
+        }
+
         found.setNomeTorneo(torneoDTO.nomeTorneo());
         found.setDataInizio(torneoDTO.dataInizio());
         found.setDataFine(torneoDTO.dataFine());
         found.setNumeroMassimoPartecipanti(torneoDTO.numeroMassimoPartecipanti());
         found.setStatoTorneo(torneoDTO.statoTorneo());
-        found.setDescrizione(torneoDTO.descrizione()); // Aggiorna la descrizione
+        found.setDescrizione(torneoDTO.descrizione());
         return torneoRepository.save(found);
     }
 
-    public void deleteTorneo(UUID torneoId) {
+    public void deleteTorneo(UUID torneoId, Utente utenteAutenticato) {
         Torneo found = findById(torneoId);
+
+        // Verifica se l'utente è l'organizzatore
+        if (!found.getOrganizzatore().getId().equals(utenteAutenticato.getId())) {
+            throw new UnauthorizedException("Non sei autorizzato a eliminare questo torneo");
+        }
+
         torneoRepository.delete(found);
     }
 }
